Update file structure, add ClothOnTable test sample，ERROR
