+++ python/Dynamics/PyCloth.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../PyCommon.h"
+
+void pybind_cloth(py::module& m);
+
