--- python/PyCloth.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "PyCommon.h"
-
-void pybind_cloth(py::module& m);
