+++ python/PyFramework.cpp
@@ -1,164 +1,5 @@
 #include "PyFramework.h"
 
-template<class TNode, class ...Args>
-std::shared_ptr<TNode> create_root(SceneGraph& scene, Args&& ... args) {
-	return scene.createNewScene<TNode>(std::forward<Args>(args)...);
-}
-
-void pybind_log(py::module& m)
-{
-	//TODO: Log is updated, update the python binding as well
-// 	py::class_<Log>(m, "Log")
-// 		.def_static("set_output", &Log::setOutput)
-// 		.def_static("get_output", &Log::getOutput)
-// 		.def_static("set_level", &Log::setLevel);
-}
-
-template<typename T, DeviceType deviceType>
-void declare_array(py::module& m, std::string typestr) {
-	using Class = dyno::FArray<T, deviceType>;
-	using Parent = FBase;
-	std::string pyclass_name = std::string("Array") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("resize", &Class::resize);
-}
-
-template<typename T>
-void declare_instance(py::module& m, std::string typestr) {
-	using Class = dyno::FInstance<T>;
-	using Parent = InstanceBase;
-	std::string pyclass_name = std::string("Instance") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("connect", &Class::connect)
-		.def("disconnect", &Class::disconnect);
-}
-
-template <typename TDataType>
-void declare_calculate_norm(py::module& m, std::string typestr) {
-	using Class = dyno::CalculateNorm<TDataType>;
-	using Parent = dyno::ComputeModule;
-	std::string pyclass_name = std::string("CalculateNorm") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("in_vec", &Class::inVec, py::return_value_policy::reference)
-		.def("out_norm", &Class::outNorm, py::return_value_policy::reference);
-}
-
-template <typename TDataType>
-void declare_pointset(py::module& m, std::string typestr) {
-	using Class = dyno::PointSet<TDataType>;
-	using Parent = dyno::TopologyModule;
-	std::string pyclass_name = std::string("PointSet") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>());
-}
-
-template <typename TDataType>
-void declare_triangleSet(py::module& m, std::string typestr) {
-	using Class = dyno::TriangleSet<TDataType>;
-	using Parent = dyno::EdgeSet<TDataType>;
-	std::string pyclass_name = std::string("TriangleSet") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>());
-}
-
-template <typename TDataType>
-void declare_edgeSet(py::module& m, std::string typestr) {
-	using Class = dyno::EdgeSet<TDataType>;
-	using Parent = dyno::PointSet<TDataType>;
-	std::string pyclass_name = std::string("EdgeSet") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>());
-}
-
-template <typename TDataType>
-void declare_discrete_elements_to_triangle_set(py::module& m, std::string typestr) {
-	using Class = dyno::DiscreteElementsToTriangleSet<TDataType>;
-	using Parent = dyno::TopologyMapping;
-	std::string pyclass_name = std::string("DiscreteElementsToTriangleSet") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("in_discreteElements", &Class::inDiscreteElements, py::return_value_policy::reference)
-		.def("out_triangleSet", &Class::outTriangleSet, py::return_value_policy::reference);
-}
-
-void declare_discrete_topology_mapping(py::module& m, std::string typestr) {
-	using Class = dyno::TopologyMapping;
-	using Parent = dyno::Module;
-	std::string pyclass_name = std::string("TopologyMapping") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str());
-}
-
-//------------------------- New ------------------------------
-
-#include "SphereModel.h"
-template <typename TDataType>
-void declare_sphere_model(py::module& m, std::string typestr) {
-	using Class = dyno::SphereModel<TDataType>;
-	using Parent = dyno::ParametricModel<TDataType>;
-	std::string pyclass_name = std::string("SphereModel") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("var_scale", &Class::varScale, py::return_value_policy::reference)
-		.def("var_location", &Class::varLocation, py::return_value_policy::reference)
-		.def("state_triangleSet", &Class::stateTriangleSet, py::return_value_policy::reference);
-}
-
-template <typename TDataType>
-void declare_parametric_model(py::module& m, std::string typestr) {
-	using Class = dyno::ParametricModel<TDataType>;
-	using Parent = dyno::Node;
-	std::string pyclass_name = std::string("ParametricModel") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("var_location", &Class::varLocation, py::return_value_policy::reference)
-		.def("var_rotation", &Class::varRotation, py::return_value_policy::reference)
-		.def("var_scale", &Class::varScale, py::return_value_policy::reference);
-}
-
-#include "PlaneModel.h"
-template <typename TDataType>
-void declare_plane_model(py::module& m, std::string typestr) {
-	using Class = dyno::PlaneModel<TDataType>;
-	using Parent = dyno::ParametricModel<TDataType>;
-	std::string pyclass_name = std::string("PlaneModel") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("var_scale", &Class::varScale, py::return_value_policy::reference)
-		.def("state_triangleSet", &Class::stateTriangleSet, py::return_value_policy::reference);
-}
-
-#include "Mapping/MergeTriangleSet.h"
-template <typename TDataType>
-void declare_merge_triangle_set(py::module& m, std::string typestr) {
-	using Class = dyno::MergeTriangleSet<TDataType>;
-	using Parent = dyno::Node;
-	std::string pyclass_name = std::string("MergeTriangleSet") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("state_triangleSet", &Class::stateTriangleSet, py::return_value_policy::reference)
-		.def("in_first", &Class::inFirst, py::return_value_policy::reference)
-		.def("in_second", &Class::inSecond, py::return_value_policy::reference);
-}
-
-#include "SemiAnalyticalScheme/SemiAnalyticalSFINode.h"
-template <typename TDataType>
-void declare_semiAnalyticalSFI_node(py::module& m, std::string typestr) {
-	using Class = dyno::SemiAnalyticalSFINode<TDataType>;
-	using Parent = dyno::Node;
-	std::string pyclass_name = std::string("SemiAnalyticalSFINode") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("import_particle_systems", &Class::importParticleSystems, py::return_value_policy::reference)
-		.def("in_triangleSet", &Class::inTriangleSet, py::return_value_policy::reference);
-}
-
-//Init_static_plugin  - for example_3 WaterPouring
-#include "initializeModeling.h"
-#include "ParticleSystem/initializeParticleSystem.h"
-#include "SemiAnalyticalScheme/initializeSemiAnalyticalScheme.h"
 void declare_modeling_init_static_plugin(py::module& m, std::string typestr) {
 	using Class = dyno::ModelingInitializer;
 	using Parent = dyno::PluginEntry;
@@ -174,16 +15,26 @@ void declare_paticleSystem_init_static_plugin(py::module& m, std::string typestr
 	py::class_<Class, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
 		.def("paticleSystem_init_static_plugin", &PaticleSystem::initStaticPlugin);
 }
-/*
-void declare_semiAnalyticalScheme_init_static_plugin(py::module& m, std::string typestr) {
-	using Class = dyno::SemiAnalyticalSchemeInitializer;
-	using Parent = dyno::PluginEntry;
-	std::string pyclass_name = std::string("SemiAnalyticalSchemeInitializer" + typestr);
-	py::class_<Class, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def("semiAnalyticalScheme_init_static_plugin", &SemiAnalyticalScheme::initStaticPlugin);
-}*/
 
-//------------------------- NEW END ------------------------------
+void declare_discrete_topology_mapping(py::module& m, std::string typestr)
+{
+	using Class = dyno::TopologyMapping;
+	using Parent = dyno::Module;
+	std::string pyclass_name = std::string("TopologyMapping") + typestr;
+	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str());
+}
+
+
+void pybind_log(py::module& m)
+{
+	//TODO: Log is updated, update the python binding as well
+// 	py::class_<Log>(m, "Log")
+// 		.def_static("set_output", &Log::setOutput)
+// 		.def_static("get_output", &Log::getOutput)
+// 		.def_static("set_level", &Log::setLevel);
+}
+
+
 
 void pybind_framework(py::module& m)
 {
@@ -287,6 +138,21 @@ void pybind_framework(py::module& m)
 		.def("get_lower_bound", &SceneGraph::getLowerBound)
 		.def("add_node", static_cast<std::shared_ptr<Node>(SceneGraph::*)(std::shared_ptr<Node>)>(&SceneGraph::addNode));
 
+
+	//------------------------- New ------------------------------2024
+	py::class_<dyno::FilePath>(m, "FilePath")
+		.def(py::init<const std::string&>())
+		.def("string", &dyno::FilePath::string)
+		.def("path", &dyno::FilePath::path)
+		.def("__eq__", &dyno::FilePath::operator==)
+		.def("__ne__", &dyno::FilePath::operator!=)
+		.def("is_path", &dyno::FilePath::is_path)
+		.def("extensions", &dyno::FilePath::extensions)
+		.def("add_extension", &dyno::FilePath::add_extension)
+		.def("set_as_path", &dyno::FilePath::set_as_path)
+		.def("set_path", &dyno::FilePath::set_path);
+
+
 	declare_calculate_norm<dyno::DataType3f>(m, "3f");
 
 	declare_pointset<dyno::DataType3f>(m, "3f");
@@ -301,6 +167,7 @@ void pybind_framework(py::module& m)
 	declare_var<std::string>(m, "s");
 	declare_var<dyno::Vec3f>(m, "3f");
 	declare_var<dyno::TOrientedBox3D<Real>>(m, "TOrientedBox3D");
+	declare_var<dyno::FilePath>(m, "FilePath");
 
 	declare_array<float, DeviceType::GPU>(m, "1fD");
 	declare_array<dyno::Vec3f, DeviceType::GPU>(m, "3fD");
@@ -313,8 +180,7 @@ void pybind_framework(py::module& m)
 
 	// New
 	declare_parametric_model<dyno::DataType3f>(m, "3f");
-	declare_plane_model<dyno::DataType3f>(m, "3f");
-	declare_sphere_model<dyno::DataType3f>(m, "3f");
+
 	declare_merge_triangle_set<dyno::DataType3f>(m, "3f");
 
 	declare_semiAnalyticalSFI_node<dyno::DataType3f>(m, "3f");
