--- python/PyParticleSystem.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "PyCommon.h"
-
-void pybind_particle_system(py::module& m);
