--- python/PyQtGUI.h
@@ -1,2 +1,0 @@
-#pragma once
-void pybind_qt_gui(py::module& m);
