--- python/PyRendering.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "PyCommon.h"
-
-void pybind_rendering(py::module& m);
