+++ python/Rendering/PyQtGUI.cpp
@@ -1,18 +1,4 @@
-#include "PyCommon.h"
-
-#include "ImColorbar.h"
-#include "ImWidget.h"
-
-template<typename T>
-void declare_instance(py::module& m, std::string typestr) {
-	using Class = dyno::FInstance<T>;
-	using Parent = InstanceBase;
-	std::string pyclass_name = std::string("Instance") + typestr;
-	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str(), py::buffer_protocol(), py::dynamic_attr())
-		.def(py::init<>())
-		.def("connect", &Class::connect)
-		.def("disconnect", &Class::disconnect);
-}
+#include "PyQtGUI.h"
 
 void declare_im_colorbar(py::module& m, std::string typestr) {
 	using Class = dyno::ImColorbar;
@@ -48,6 +34,7 @@ void declare_im_widget(py::module& m, std::string typestr) {
 	std::string pyclass_name = std::string("ImWidget") + typestr;
 	py::class_<Class, Parent, std::shared_ptr<Class>>(m, pyclass_name.c_str());
 }
+
 
 void pybind_qt_gui(py::module& m)
 {
